feat(config): merge BinConfig & Config

+++ crates/daemon/src/lib/config.rs
@@ -13,12 +13,24 @@ pub struct Config {
 	#[serde(default = "Config::default_feed_path")]
 	/// Directory to store all feed related files in
 	pub feed_path: String,
+	#[serde(default = "Config::latest_version")]
+	/// Config version
+	pub version:    usize,
+	#[serde(default = "Config::default_logpath")]
+	/// Directory to place log files
+	pub log_path:   String,
+	#[serde(flatten)]
+	/// serde unkown fields
+	pub extra:      HashMap<String, serde_yaml::Value>,
 }
 
 impl Default for Config {
 	fn default() -> Self {
 		return Config {
 			feed_path: Self::default_feed_path(),
+			version:    Self::latest_version(),
+			log_path:   Self::default_logpath(),
+			extra:      HashMap::new(),
 		};
 	}
 }
@@ -27,37 +39,7 @@ impl Config {
 	pub fn default_feed_path() -> String {
 		return "./feed".to_string();
 	}
-}
 
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct BinConfig {
-	#[serde(default = "BinConfig::latest_version")]
-	/// Config version
-	pub version:    usize,
-	#[serde(flatten)]
-	/// Config to be passed to the lib
-	pub lib_config: Config,
-	#[serde(default = "BinConfig::default_logpath")]
-	/// Directory to place log files
-	pub log_path:   String,
-	#[serde(flatten)]
-	/// serde unkown fields
-	pub extra:      HashMap<String, serde_yaml::Value>,
-}
-
-impl Default for BinConfig {
-	fn default() -> Self {
-		return Self {
-			version:    Self::latest_version(),
-			log_path:   Self::default_logpath(),
-			lib_config: Config::default(),
-			extra:      HashMap::new(),
-		};
-	}
-}
-
-impl BinConfig {
 	pub fn latest_version() -> usize {
 		return 0;
 	}
@@ -68,27 +50,34 @@ impl BinConfig {
 
 	pub fn from_cli_matches(cli_matches: &clap::ArgMatches) -> anyhow::Result<Self> {
 		let config_path = PathBuf::from(&cli_matches.value_of("config").unwrap());
-		debug!("Config Path is {:?}", &config_path);
 
-		let config: BinConfig = if Path::new(&config_path).exists() {
-			debug!("Config File Exists");
-			let config_raw = File::open(&config_path).context("Opening config path for reading Failed")?;
-			serde_yaml::from_reader(&config_raw).context("Couldnt read config")?
-		} else {
-			debug!("Config File does not exist");
-			let config = BinConfig::default();
-			let write = File::create(&config_path).context("Opening config path for writing Failed")?;
-			serde_yaml::to_writer(&write, &config).context("Writing default config failed")?;
-			config
-		};
-
-		debug!("Config File's Content: {:#?}", &config);
+		let config = Self::read_or_create_config(&config_path)?;
 
 		if !config.extra.is_empty() {
 			for key in config.extra.keys() {
 				warn!("Unkown key in config: \"{}\"", &key);
 			}
 		}
+
+		return Ok(config);
+	}
+
+	fn read_or_create_config(config_path: &PathBuf) -> anyhow::Result<Self> {
+		debug!("Config Path is {:?}", &config_path);
+
+		let config: Self;
+		if Path::new(&config_path).exists() {
+			debug!("Config File Exists");
+			let config_raw = File::open(&config_path).context("Opening config path for reading Failed")?;
+			config = serde_yaml::from_reader(&config_raw).context("Couldnt read config")?;
+		} else {
+			debug!("Config File does not exist");
+			config = Config::default();
+			let write = File::create(&config_path).context("Opening config path for writing Failed")?;
+			serde_yaml::to_writer(&write, &config).context("Writing default config failed")?; 
+		};
+
+		debug!("Config File's Content: {:#?}", &config);
 
 		return Ok(config);
 	}
